parts done

+++ Assets/App/Scripts/GameScene/Gameplay Management/Block Management/BlockInteractionManager.cs
@@ -3,6 +3,7 @@ using App.GameScene.Blocks;
 using App.GameScene.Gameplay_Management.Input_Management;
 using App.GameScene.Visualization;
 using DG.Tweening;
+using Unity.VisualScripting;
 using UnityEngine;
 
 namespace App.GameScene.Gameplay_Management.Block_Management
@@ -10,9 +11,11 @@ namespace App.GameScene.Gameplay_Management.Block_Management
     public class BlockInteractionManager : Manager
     {
         private readonly List<Block> _blocks = new();
-        [SerializeReference] private CameraManager cameraManager;
         
+        [SerializeReference] private CameraManager cameraManager;
         private Rect _cameraSize;
+        
+        [SerializeField] private Part partPrefab;
 
         public override void Init()
         {
@@ -47,7 +50,7 @@ namespace App.GameScene.Gameplay_Management.Block_Management
                         new Rect((Vector2)block.transform.position - new Vector2(block.Radius, block.Radius),
                             new Vector2(block.Radius * 2, block.Radius * 2)))) continue;
                 
-                //block.OnMiss();
+                block.OnMiss();
                 DeleteBlock(i);
                 i--;
             }
@@ -58,11 +61,60 @@ namespace App.GameScene.Gameplay_Management.Block_Management
             for (var i = 0; i < _blocks.Count; i++)
             {
                 var block = _blocks[i];
+                if (!block.IsInteractable) continue;
                 if (!(DistanceToSegment(deathLine, block) <= block.Radius)) continue;
-                //block.OnHit();
+
+                if (block.IsHalfable)
+                {
+                    CreateAndSetupParts(deathLine, block);
+                }
+                
+                block.OnHit();
                 DeleteBlock(i);
                 i--;
             }
+        }
+
+        private void CreateAndSetupParts(DeathLine deathLine, Block block)
+        {
+            Vector2 blockPosition = block.transform.position;
+            
+            var sprite = block.spriteRenderer.sprite;
+            var pixelSpriteSize = sprite.rect.size;
+            var spriteSize = sprite.bounds.size;
+                    
+            var deathLineDirection = (deathLine.To - deathLine.From).normalized;
+            var perp = Vector2.Perpendicular(deathLineDirection);
+            
+            var rotatedDeathLineDirection = Quaternion.Euler(0, 0, 90) * deathLineDirection;
+            var rotation = Quaternion.LookRotation(Vector3.forward, rotatedDeathLineDirection);
+
+            var firstPartPosition = blockPosition + spriteSize.y / 4f * perp;
+            var secondPartPosition = blockPosition - spriteSize.y / 4f * perp;
+                    
+            var firstPart = Instantiate(partPrefab, firstPartPosition, rotation);
+            var secondPart = Instantiate(partPrefab, secondPartPosition, rotation);
+            
+            var firstPartRect = new Rect(0, pixelSpriteSize.y / 2f, pixelSpriteSize.x, pixelSpriteSize.y / 2f);
+            var secondPartRect = new Rect(0, 0, pixelSpriteSize.x, pixelSpriteSize.y / 2f);
+                    
+            firstPart.SetSprite(Sprite.Create(sprite.texture,
+                firstPartRect,
+                new Vector2(0.5f, 0.5f)));
+                    
+            secondPart.SetSprite(Sprite.Create(sprite.texture,
+                secondPartRect, 
+                new Vector2(0.5f, 0.5f)));
+                    
+                    
+            firstPart.ThrowItself(firstPartPosition, deathLine.Speed * deathLineDirection / 5f + perp * deathLine.Speed / 10f);
+            secondPart.ThrowItself(secondPartPosition, deathLine.Speed * deathLineDirection / 5f - perp * deathLine.Speed / 10f);
+
+            firstPart.transform.DORotate(rotation.eulerAngles + new Vector3(0, 0, 360), 5, RotateMode.FastBeyond360).SetLoops(-1);
+            secondPart.transform.DORotate(rotation.eulerAngles + new Vector3(0, 0, -360), 5, RotateMode.FastBeyond360).SetLoops(-1);
+                    
+            AddBlock(firstPart);
+            AddBlock(secondPart);
         }
         
         private static float DistanceToSegment(DeathLine deathLine, Component block)
